upload with bucket

+++ app/Services/Panel/Course/CourseFileService.php
@@ -2,42 +2,21 @@
 
 namespace App\Services\Panel\Course;
 
-use App\Models\Course;
 use App\Models\CourseFile;
-use App\Models\Lesson;
-use App\Models\LessonFile;
-use App\Models\LessonSampleWork;
-use App\Models\Level;
-use App\Models\LevelCategory;
-use App\Models\SecretKey;
-use App\Models\UserLevelCategory;
-use App\Services\UploaderService;
-use App\Services\Panel\Message\MessageService;
-use App\Services\SecretKeyService;
-use App\ViewModel\Lesson\LessonContent\SaveCourseFileViewModel;
-use App\ViewModel\Lesson\LessonContent\SaveCourseViewModel;
-use App\ViewModel\Lesson\LessonContent\NewSampleWorkViewModel;
-use App\ViewModel\Lesson\LessonContent\SaveLessonFileViewModel;
-use App\ViewModel\Lesson\LessonContent\SaveLessonViewModel;
-use App\ViewModel\Message\SendMessageViewModel;
-use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\File;
-use Illuminate\Support\Facades\Response;
-use Illuminate\Support\Facades\Storage;
+use App\Services\bucket\BucketService;
+use App\ViewModel\Course\SaveCourseFileViewModel;
 use Illuminate\Support\Str;
-use phpDocumentor\Reflection\Types\True_;
-use Ramsey\Uuid\Guid\Guid;
 
 class CourseFileService
 {
 
-    private UploaderService $uploaderService;
+    private BucketService $bucketService;
 
     public function __construct()
     {
         ini_set('max_execution_time', -1);
 
-        $this->uploaderService = new UploaderService();
+        $this->bucketService = new BucketService();
     }
 
 
@@ -45,19 +24,21 @@ class CourseFileService
         $courseFile = new CourseFile();
         $courseFile->course_id = $model->getCourseId();
         $courseFile->title = $model->getTitle();
-        $destinationAddress = "course".DIRECTORY_SEPARATOR .$model->getCourseId();
+        $destinationAddress = "course"."/" .$model->getCourseId()."/".$model->getFile()['name'];
         $courseFile->secret_key = Str::uuid();
-        $result = $this->uploaderService->saveFile($model->getFile(), $destinationAddress);
-        $courseFile->file_path = $result['file_name'];
-        $courseFile->postfix = ".".$result['postfix'];
+        $result = $this->bucketService->uploadPartOfFile($model->getFile(), $destinationAddress);
+        $courseFile->file_path = $model->getFile()['name'];
+        $courseFile->postfix = $model->getFile()['type'];
 
         $courseFile->save();
     }
 
     public function DeleteCourseFile($id){
         $courseFile = CourseFile::find($id);
-        $fullPath = "course".DIRECTORY_SEPARATOR.$courseFile->lesson_id;
-        $this->uploaderService->unlink($fullPath, $courseFile->file_path);
+        if(!$courseFile)
+            abort(404);
+        $fullPath = "course"."/".$courseFile->lesson_id."/".$courseFile->file_path;
+        $this->bucketService->Delete($fullPath);
         $courseFile->delete();
     }
